Feat(front) : 프로필 수정 페이지 기능 구현 (S14P11D102-199)

+++ FrontEnd/app/src/main/java/com/ssafy/closetory/dto/EditProfileUpdateData.kt
@@ -3,7 +3,7 @@ package com.ssafy.closetory.dto
 data class EditProfileUpdateData(
     val nickname: String,
     val gender: String,
-    val height: Int,
-    val weight: Int,
+    val height: Short,
+    val weight: Short,
     val alarmEnabled: Boolean
 )
